requirements.txt
kafka-python==2.0.2
python-dotenv==0.15.0
